# remote_write: Only use a slice when we have a hash conflict