#### PR Description

Syncs code from the upstream memSeries implemetation (https://github.com/prometheus/prometheus/blob/main/tsdb/agent/series.go) to our version to further reduce the memory footprint of the WAL. The idea is that we do not often have hash conflicts for series and having a lot single entry slices is a lot of unnecessary overhead. Only using a slice when there's a conflict should reduce this by a fair amount.

bench results for reference

```
benchstat series-baseline.txt series-new.txt
goos: darwin
goarch: arm64
pkg: github.com/grafana/alloy/internal/static/metrics/wal
cpu: Apple M3 Pro
                │ series-baseline.txt │           series-new.txt            │
                │       sec/op        │    sec/op     vs base               │
CreateSeries-11           539.8n ± 9%   455.0n ± 11%  -15.70% (p=0.004 n=6)

                │ series-baseline.txt │           series-new.txt           │
                │        B/op         │    B/op      vs base               │
CreateSeries-11            264.0 ± 9%   149.5 ± 33%  -43.37% (p=0.002 n=6)

                │ series-baseline.txt │          series-new.txt           │
                │      allocs/op      │ allocs/op   vs base               │
CreateSeries-11            2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.002 n=6)

```

#### Notes to the Reviewer
After this our there's only a few differences between our implementation and upstream which I will attempt to sync (the WAL bugfix being the most important one). We might be able to switch to using upstream directly here instead but it's not as high of a priority. 

#### PR Checklist

- [x] CHANGELOG.md updated
- [x] Tests updated